Move account dashboard components to account route
